LayoutTests/http/tests/contentextensions/block-cookies-in-ping-expected.txt

This test sets a cookie, then follows a link with a ping attribute where the ping URL matches a 'block-cookie' rule.   

--------
Frame: 'link_frame'
--------
Link with ping was clicked.

--------
Frame: 'result_frame'
--------
Ping received.
No cookies in ping.
